working login

+++ jTLC/TKConstants.h
@@ -10,7 +10,7 @@
 #define TKLoginURL              [NSURL URLWithString:@"https://mytlc.bestbuy.com/etm/login.jsp"]
 #define TKAvailabilityChangeURL [NSURL URLWithString:@"https://mytlc.bestbuy.com/etm/messaging/etmBOView.jsp?msg_id=null&busObjTyp_id=20051&folder_id=null&newOld=null&parentID=1000000442&selectedTocID=1000000303&parentID=1000000442"]
 #define TKCalendarURL           [NSURL URLWithString:@"https://mytlc.bestbuy.com/etm/time/timesheet/etmTnsMonth.jsp"]
-#define TKStoreDetailURL(num)   [NSURL URLWithString:[NSString stringWithFormat:@"http://stores.bestbuy.com/%li/details", (num)]]
+#define TKStoreDetailURL(num)   [NSURL URLWithString:[NSString stringWithFormat:@"http://stores.bestbuy.com/%li/details", (long)(num)]]
 
 typedef struct {
     NSInteger hour;
